+++ examples/cuda/reduction.py
@@ -14,6 +14,7 @@ def tune():
     tune_params["use_shuffle"] = [0, 1]
     tune_params["vector"] = [2**i for i in range(3)]
     tune_params["num_blocks"] = [2**i for i in range(5,16)]
+    tune_params["loop_unroll_factor_0"] = [0, 1, 8, 16, 32, 64]
 
     problem_size = "num_blocks"
     size = 800000000