+++ kernel_tuner/interface.py
@@ -319,7 +319,8 @@ _tuning_options = Options([
     ("cache", ("""filename for caching/logging benchmarked instances
         filename uses suffix ".json"
         if the file exists it is read and tuning continues from this file
-        """, "string"))
+        """, "string")),
+    ("metrics", ("specifies user-defined metrics", "OrderedDict"))
     ])
 
 _device_options = Options([
@@ -368,7 +369,7 @@ def tune_kernel(kernel_name, kernel_string, problem_size, arguments,
                 compiler=None, compiler_options=None, log=None,
                 iterations=7, block_size_names=None, quiet=False, strategy=None, strategy_options=None,
                 scheduler=None,
-                cache=None):
+                cache=None, metrics=None):
 
     if log:
         logging.basicConfig(filename=kernel_name + datetime.now().strftime('%Y%m%d-%H:%M:%S') + '.log', level=log)