+++ test/test_util_functions.py
@@ -126,8 +126,9 @@ def test_prepare_kernel_string():
     params = dict()
     params["is"] = 8
 
-    _, output = prepare_kernel_string("this", kernel, params, (3,7), (1,2,3), block_size_names)
-    expected = "#define is 8\n" \
+    _, output = prepare_kernel_string("this", kernel, params, (3,7), (1,2,3), block_size_names, "")
+    expected = "#define kernel_tuner 1\n" \
+               "#define is 8\n" \
                "#define block_size_z 3\n" \
                "#define block_size_y 2\n" \
                "#define block_size_x 1\n" \
@@ -487,3 +488,38 @@ def test_process_cache():
     finally:
         delete_temp_file(cache)
         #pass
+
+
+def test_process_metrics():
+    params = {"x": 15, "b": 12}
+    metrics = OrderedDict()
+    metrics["y"] = lambda p : p["x"]
+
+    #test if lambda function is correctly evaluated
+    params = process_metrics(params, metrics)
+    assert params["y"] == params["x"]
+
+    #test if we can do the same with a string
+    params = {"x": 15, "b": 12}
+    metrics["y"] = "x"
+    params = process_metrics(params, metrics)
+    assert params["y"] == params["x"]
+
+    #test if composability works correctly
+    params = {"x": 15, "b": 12}
+    metrics = OrderedDict()
+    metrics["y"] = "x"
+    metrics["z"] = "y"
+    params = process_metrics(params, metrics)
+    assert params["z"] == params["x"]
+
+    #test ValueError is raised when metrics is not an OrderedDict
+    with pytest.raises(ValueError):
+        params = process_metrics(params, {})
+
+    #test ValueError is raised when b already exists in params
+    params = {"x": 15, "b": 12}
+    metrics = OrderedDict()
+    metrics["b"] = "x"
+    with pytest.raises(ValueError):
+        params = process_metrics(params, metrics)