+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Added
+- support loop unrolling using params that start with loop_unroll_factor
+- always insert "define kernel_tuner" to allow preprocessor ifdef kernel_tuner
+- support for user-defined metrics
 
 ## [0.3.1] - 2020-06-11
 ### Added